eightfold-company-assistant/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.40.0
pydantic